# BackEnd/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TsunamiIT Manager Project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/scrumTeam": {
            "post": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Create Scrum Team",
                "operationId": "85bca1d9da6059332f54e7b9fe70f405",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Team Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/deleteScrumTeam/{idScrumTeam}": {
            "delete": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Delete a Scrum Team",
                "operationId": "199886f0b70b693b7680863950640913",
                "parameters": [
                    {
                        "name": "idScrumTeam",
                        "in": "path",
                        "description": "ScrumTeam ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "apiAuth": [
                            ""
                        ]
                    }
                ]
            }
        },
...
        "/api/getScrumTeamsByProductOwner/{idProductOwner}": {
            "get": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Returns all ScrumTeams by product owner Id",
                "operationId": "5655734c40c6ff934ff7a211f98c2fb7",
                "parameters": [
                    {
                        "name": "idProductOwner",
                        "in": "path",
                        "description": "Product Owner ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "No team found"
                    }
                }
            }
        },
        "/api/getScrumTeamById/{idScrumTeam}": {
            "get": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Returns the ScrumTeams by Id",
                "operationId": "8ef44f7710e089a6f260eb652344d0a0",
                "parameters": [
                    {
                        "name": "idScrumTeam",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "No team found"
                    }
                }
            }
        },
        "/api/scrumteam/{idScrumTeam}": {
            "get": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Returns all Collaborators",
                "operationId": "d6c617478598a3f779b57b4f8a16afad",
                "parameters": [
                    {
                        "name": "idScrumTeam",
                        "in": "path",
                        "description": "ScrumTeam ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    }
                }
            }
        },
        "/api/currentscrumteam/project/{idProject}": {
            "get": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Returns the scrum Team that works currently on a project",
                "operationId": "6a014a61fd65ed32015c7377ceffad8c",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    },
                    "404": {
                        "description": "no Project or available current team working on that project found!"
                    }
                }
            }
        },
        "/api/scrumteam/project/{idProject}": {
            "get": {
                "tags": [
                    "ScrumTeam"
                ],
                "summary": "Returns all the scrum Teams that worked on a project",
                "operationId": "e605ab1351a950082aadb869e2910e41",
                "parameters": [
                    {
                        "name": "idProject",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation !"
                    }
                }
            }
        },
...
    }
}